+++ tests/bootstrap.php
@@ -24,6 +24,8 @@ require_once $fixtures . '/class-wp-posts.php';
 require_once $fixtures . '/wp-functions.php';
 
 defined('ABSPATH') || define('ABSPATH', $fixtures . '/');
+defined('WP_CONTENT_DIR') || define('WP_CONTENT_DIR', dirname(__DIR__) . '/wp-content');
+defined('WP_PLUGIN_DIR') || define('WP_PLUGIN_DIR', WP_CONTENT_DIR . '/plugins');
 defined('WP_NERVE_VERSION') || define('WP_NERVE_VERSION', '0.1.0-alpha.3');
 defined('WP_NERVE_FILE') || define('WP_NERVE_FILE', dirname(__DIR__) . '/wp-nerve.php');
 defined('WP_NERVE_PATH') || define('WP_NERVE_PATH', dirname(__DIR__) . '/');